Syntax highlighting in editor works.
Found that source.asciidoc had to be added to  list  of  grammars detected
TO DO  add this to default list somewhere in the program and remove some (or all) of  markdown versions.
THE NEXT  commit on 20240405 c 1240 DOES NOT WORK OK - there is a problem wi  save .adoc.html  to do with styles.