Last version before a copy  made to  IansascidocPreviewFrommarkdownpreviewnomd
where all the bits no needed by asciidoctor will be removed.
This version seems to work well.
Except that save as html works for everything but not for asciidoc_syntax_Long.adoc. If sections of that file are removed
it works fine. There is something in that file that prevents writing to disk - the console preview of the file seems OK.